+++ Tyuiu.AkhmetovRR.Sprint1.Task5.V4.Test/DataServiceTest.cs
@@ -9,7 +9,7 @@ namespace Tyuiu.AkhmetovRR.Sprint1.Task5.V4.Test
         {
             double k = 13257;
             DataService ds = new DataService();
-            double result = ds.Calculate(k);
+            int result = ds.Calculate(k);
             Assert.AreEqual(Math.Round(k / 3600) - 1, result);
         }
     }
